Extract Three.js typings to `threeJsCompat`
