+++ DevToolkt/jsApiCompat/src/jsMain/kotlin/dev/toolkt/js/utils.kt
@@ -1,4 +1,4 @@
-package org.w3c.dom.extra
+package dev.toolkt.js
 
 @Suppress("NOTHING_TO_INLINE")
 inline fun jsObject(): dynamic = js("({})")
