+++ cupLayoutTool/src/jsMain/kotlin/diy/lingerie/web_tool_3d/presentation/MyBezierMesh.kt
@@ -5,9 +5,9 @@ import dev.toolkt.geometry.Point3D
 import dev.toolkt.reactive.cell.Cell
 import diy.lingerie.web_tool_3d.UserBezierMesh
 import diy.lingerie.web_tool_3d.application_state.InteractionState
-import three.MeshBasicMaterialParams
-import three.MeshLambertMaterialParams
-import three.THREE
+import dev.toolkt.js.threejs.MeshBasicMaterialParams
+import dev.toolkt.js.threejs.MeshLambertMaterialParams
+import dev.toolkt.js.threejs.THREE
 
 class MyBezierMesh(
     val root: THREE.Object3D,
