+++ cupLayoutTool/src/jsMain/kotlin/diy/lingerie/web_tool_3d/presentation/MyObjectUserData.kt
@@ -1,7 +1,7 @@
 package diy.lingerie.web_tool_3d.presentation
 
 import diy.lingerie.web_tool_3d.UserBezierMesh
-import three.THREE
+import dev.toolkt.js.threejs.THREE
 
 sealed class MyObjectUserData {
     data class HandleBallUserData(
