+++ cupLayoutTool/src/jsMain/kotlin/diy/lingerie/web_tool_3d/presentation/flatFabricPieceUtils.kt
@@ -4,8 +4,8 @@ import dev.toolkt.dom.pure.PureColor
 import dev.toolkt.geometry.Point3D
 import dev.toolkt.reactive.cell.Cell
 import diy.lingerie.web_tool_3d.application_state.FlatFabricPiece
-import three.MeshLambertMaterialParams
-import three.THREE
+import dev.toolkt.js.threejs.MeshLambertMaterialParams
+import dev.toolkt.js.threejs.THREE
 
 private const val fabricParticleRadius = 0.75
