+++ cupLayoutTool/src/jsMain/kotlin/diy/lingerie/web_tool_3d/presentation/threeReactiveUtils.kt
@@ -13,9 +13,9 @@ import dev.toolkt.reactive.cell.Cell
 import dev.toolkt.reactive.cell.MutableCell
 import kotlinx.browser.window
 import org.w3c.dom.HTMLCanvasElement
-import three.THREE
-import three.THREE.Object3D
-import three.WebGLRendererParams
+import dev.toolkt.js.threejs.THREE
+import dev.toolkt.js.threejs.THREE.Object3D
+import dev.toolkt.js.threejs.WebGLRendererParams
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.milliseconds
